Introducing templated to time stepper
