--- src/time_steppers/ac_time_stepper_for_odes.cpp
@@ -1,23 +1,0 @@
-#include "ac_time_stepper_for_odes.h"
-
-namespace scicellxx
-{
-
- // ===================================================================
- // Empty constructor
- // ===================================================================
- ACTimeStepperForODEs::ACTimeStepperForODEs()
-  : ACTimeStepper()
- {
-  
- }
- 
- // ===================================================================
- // Empty destructor
- // ===================================================================
- ACTimeStepperForODEs::~ACTimeStepperForODEs()
- { 
-
- }
- 
-}
