+++ src/time_steppers/cc_euler_method.h
@@ -1,51 +1,7 @@
-#ifndef CCEULERMETHOD_H
-#define CCEULERMETHOD_H
-
-#include "ac_time_stepper_for_odes.h"
-
-namespace scicellxx
-{
-
- /// @class CCEulerMethod cc_euler_method.h
- /// This class implements Euler's method to integrate ODE's
- class CCEulerMethod : public virtual ACTimeStepperForODEs
- {
- 
- public:
-
-  /// Constructor
-  CCEulerMethod();
-  
-  /// Empty destructor
-  virtual ~CCEulerMethod();
-  
-  /// Applies Eulers method to the given odes from the current time
-  /// "t" to the time "t+h". The values of u at time t+h will be
-  /// stored at index k (default k = 0).
-  void time_step(ACODEs &odes, const Real h, const Real t,
-                 CCData &u, const unsigned k = 0);
-  
- protected:
-  
-  /// Copy constructor (we do not want this class to be
-  /// copiable). Check
-  /// http://www.learncpp.com/cpp-tutorial/912-shallow-vs-deep-copying/
-  CCEulerMethod(const CCEulerMethod &copy)
-   : ACTimeStepperForODEs()
-   {
-    BrokenCopy::broken_copy("CCEulerMethod");
-   }
-  
-  /// Assignment operator (we do not want this class to be
-  /// copiable. Check
-  /// http://www.learncpp.com/cpp-tutorial/912-shallow-vs-deep-copying/
-  void operator=(const CCEulerMethod &copy)
-   {
-    BrokenCopy::broken_assign("CCEulerMethod");
-   }
-
- };
-
-}
- 
-#endif // #ifndef CCEULERMETHOD_H
+// Include header and tpl.cpp files implementing templates. We may use
+// this file to force the instantiation of specific templates
+#ifndef CC_EULERMETHOD_H
+#define CC_EULERMETHOD_H
+#include "cc_euler_method.tpl.h"
+#include "cc_euler_method.tpl.cpp"
+#endif // #ifndef CC_EULERMETHOD_H
